chore: add initial worker-bridge test
